# Refactor: (Performance) `register_graphql_interfaces` no longer used in `register_block_types`