--- a/.changeset/thin-insects-sparkle.md
+++ b/.changeset/thin-insects-sparkle.md
@@ -1,0 +1,5 @@
+---
+"@wpengine/wp-graphql-content-blocks": patch
+---
+
+Refactored `register_block_types` to remove usages of `register_graphql_interfaces_to_types` to improve performance.