--- a/.changeset/thin-insects-sparkle.md
+++ b/.changeset/thin-insects-sparkle.md
@@ -3,3 +3,4 @@
 ---
 
 Refactored `register_block_types` to remove usages of `register_graphql_interfaces_to_types` to improve performance.
+**MINOR BREAKING** Removed `Anchor::register_to_block` static method.