--- a/includes/Field/BlockSupports/Anchor.php
+++ b/includes/Field/BlockSupports/Anchor.php
@@ -9,6 +9,7 @@ namespace WPGraphQL\ContentBlocks\Field\BlockSupports;
 
 use WP_Block_Type;
 use WPGraphQL\ContentBlocks\Utilities\DOMHelpers;
+use WPGraphQL\ContentBlocks\Utilities\WPGraphQLHelpers;
 
 /**
  * Class - Anchor
@@ -64,5 +65,20 @@ class Anchor {
 	 */
 	public static function get_block_attributes_interfaces( $existing, \WP_Block_Type $block_spec ): array {
 		return self::get_block_interfaces( $existing, $block_spec );
+	}
+
+	/**
+	 * Registers an Anchor field on a block if it supports it.
+	 *
+	 * @param \WP_Block_Type $block_spec The block type to register the anchor field against.
+	 * @return void
+	 *
+	 * @deprecated 1.1.4 No longer used by internal code and not recommended.
+	 */
+	public static function register_to_block( \WP_Block_Type $block_spec ): void {
+		if ( isset( $block_spec->supports['anchor'] ) && true === $block_spec->supports['anchor'] ) {
+			register_graphql_interfaces_to_types( 'BlockWithSupportsAnchor', array( WPGraphQLHelpers::format_type_name( $block_spec->name ) . 'Attributes' ) );
+			register_graphql_interfaces_to_types( 'BlockWithSupportsAnchor', array( WPGraphQLHelpers::format_type_name( $block_spec->name ) ) );
+		}
 	}
 }