@theodesp 2023-08-15
https://github.com/wpengine/wp-graphql-content-blocks/issues/134