@theodesp 2023-08-16
@jasonbahl @justlevine It looks like `eagerLoadType` does impact performance:

Its 62% slower when enabled and 47% slower when disabled.