@justlevine 2023-08-16
> @jasonbahl @justlevine It looks like `eagerLoadType` does impact performance:
> 
> Its 62% slower when enabled to 47% slower when disabled.

What exactly is the reason for turning on `eagerlyLoadType`? Is it just [for the IDE](https://github.com/wp-graphql/wp-graphql/issues/2598) or is it required for valid   resolution of the type?