@theodesp 2023-08-16
Hey @justlevine  the IDE wont' show any documentation about the Block types:

<img width="1558" alt="Screenshot 2023-08-16 at 12 25 41" src="https://github.com/wpengine/wp-graphql-content-blocks/assets/328805/72f22021-fdea-4a84-a252-5249016dfb3a">

This is not ideal.